Fix CI (#92)
This PR fixes the CI. The problem was that node was updated to version 18 on ubuntu runners. We don't really want to update webpack so we have to use Node Version 16 so we are able to compile the assets.

Also introduces `.node-version` and `.ruby-version` files to be consisted with the updated `ci.yml` file.