Do not use `permanent` when reading from cookies (#91)
When we want to store a permanent cookie, we have to use the `permanent` accessor. When reading, this is not needed. See [documentation (https://api.rubyonrails.org/classes/ActionDispatch/Cookies/ChainedCookieJars.html#method-i-permanent) :

> This jar is only meant for writing. You'll read permanent cookies through the regular accessor.

When we want to check for a cookie existence, instead of doing:

`cookies.encrypted[:remember_token].present?`

we can do just:

`cookies[:remember_token]`